Media-source backed elements block load event; cause web-platform-test flakiness https://bugs.webkit.org/show_bug.cgi?id=161881
Reviewed by Eric Carlson.

Source/WebCore:

Test: media/media-source/media-source-delaying-load-event.html

The MSE specification has added an explicit step to their "attaching to media element"
algorithm which tells the media element to stop delaying the load event. And indeed,
the HTMLMediaElement blocks the load event when a MediaSource is attached but its data
is never loaded.

* Modules/mediasource/MediaSource.cpp:
(WebCore::MediaSource::setPrivateAndOpen):
* html/HTMLMediaElement.h:

LayoutTests:

* media/media-source/media-source-delaying-load-event-expected.txt: Added.
* media/media-source/media-source-delaying-load-event.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@205863 268f45cc-cd09-0410-ab3c-d52691b4dbfc