Only override URLParser::enabled if NSUserDefaults object exists https://bugs.webkit.org/show_bug.cgi?id=162131
Patch by Alex Christensen <achristensen@webkit.org> on 2016-09-19
Reviewed by Sam Weinig.

* Shared/Cocoa/WebKit2InitializeCocoa.mm:
(WebKit::platformInitializeWebKit2):
* Shared/WebKit2Initialize.cpp:
When I enable URLParser by default, I will want the default to only be overwritten if there is a
NSUserDefaults key for URLParserEnabled.  Right now it is setting URLParser::enabled to false always,
which isn't a problem because the default is also false.  When the default is true, that will overwrite the default.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206097 268f45cc-cd09-0410-ab3c-d52691b4dbfc