Optimize URLParser https://bugs.webkit.org/show_bug.cgi?id=162338
Reviewed by Tim Horton.

No change in behavior. Just a marginal performance improvement.

* platform/URLParser.cpp:
(WebCore::isValidSchemeCharacter):
Added to reduce branches when parsing the scheme of a URL.  Now there is a table lookup instead of many range checks.
(WebCore::URLParser::parse):
Use appendVector, which uses memcpy.
(WebCore::URLParser::parseHostAndPort):
If we know that we are parsing the serialized output of a valid URL, 
we do not need to convert host characters to lower case.
They are already lower case.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206223 268f45cc-cd09-0410-ab3c-d52691b4dbfc