Fullscreen controls inoperative in WebKitLegacy web views https://bugs.webkit.org/show_bug.cgi?id=162374
Reviewed by Eric Carlson.

After r205365, the WebPlaybackSessionInterfaceAVKit was now created after the
WebPlaybackSessionModel which fed it data, so it no longer received the burst of data upon
creation. Instead, it should have always asked the model for its cached data as soon as it
was connected to set up its inital state.

* platform/ios/WebPlaybackSessionInterfaceAVKit.mm:
(WebCore::WebPlaybackSessionInterfaceAVKit::WebPlaybackSessionInterfaceAVKit):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206272 268f45cc-cd09-0410-ab3c-d52691b4dbfc