REGRESSION(r204163): [GTK] API test /webkit2/WebKitSecurityManager/file-xhr is failing https://bugs.webkit.org/show_bug.cgi?id=161607
Reviewed by Michael Catanzaro.

Since r204163, in case of cross origin error when loading the XHR, the load fails and an error message is shown
in the console, but there isn't any js expcetion, which is what we were checking. Now, we check that there
aren't exceptions, and that the error message is sent to the console.

* Scripts/run-gtk-tests:
(TestRunner):
* TestWebKitAPI/Tests/WebKit2Gtk/TestWebKitWebContext.cpp:
(consoleMessageReceivedCallback):
(testWebContextSecurityFileXHR):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206304 268f45cc-cd09-0410-ab3c-d52691b4dbfc