Use topVMEntryFrame to determine whether to skip the re-throw of a simulated throw. https://bugs.webkit.org/show_bug.cgi?id=162793
Reviewed by Saam Barati.

Change the ThrowScope destructor to use topVMEntryFrame (instead of topCallFrame)
in the determination of whether to skip the re-throw of a simulated throw.  This
is needed because the topCallFrame is not updated in operationConstructArityCheck()
(and does not need to be), whereas topVMEntryFrame is always updated properly.
Hence, we should just switch to using the more reliable topVMEntryFrame instead.

This issue was discovered by existing JSC tests when exception check validation
is enabled.

* runtime/ThrowScope.cpp:
(JSC::ThrowScope::~ThrowScope):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206643 268f45cc-cd09-0410-ab3c-d52691b4dbfc