FileSaver.js does not work in WebKit https://bugs.webkit.org/show_bug.cgi?id=162788
Reviewed by Sam Weinig.

Source/WebCore:

FileSaver.js does not work in WebKit:
- https://eligrey.com/demos/FileSaver.js/

It works in Firefox and Chrome, but in WebKit, we were getting a
"Synthetic clicks on anchors that have a download attribute are
ignored." warning. We were too strict in restricting synthetic clicks.
We now allow synthetic clicks as long as they are triggered by a user
gesture.

Test: fast/dom/HTMLAnchorElement/anchor-download-user-triggered-synthetic-click.html

* html/HTMLAnchorElement.cpp:
(WebCore::HTMLAnchorElement::handleClick):

LayoutTests:

Add layout test coverage.

* fast/dom/HTMLAnchorElement/anchor-download-synthetic-click-expected.txt:
* fast/dom/HTMLAnchorElement/anchor-download-user-triggered-synthetic-click-expected.txt: Added.
* fast/dom/HTMLAnchorElement/anchor-download-user-triggered-synthetic-click.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206659 268f45cc-cd09-0410-ab3c-d52691b4dbfc