[mac-wk1] LayoutTest http/tests/security/cached-cross-origin-preloading-css-stylesheet.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=162791
Patch by Youenn Fablet <youenn@apple.com> on 2016-10-04
Reviewed by Alex Christensen.

Flackiness was due to one preload sometimes arriving after the test was finished.
This causes the issue of not having the preload actually being checked against CORS and the console error was not displayed.
Making the test wait 1 second after all loads finish to ensure the preload will also finishes.

Cleaning the test and improving it by:
- making each load to take 1 second (so that preload is not finished when actual load happens
- adding some comments explaining each subtest
- printing DONE at the end of all loads.

* http/tests/security/cached-cross-origin-preloading-css-stylesheet-expected.txt:
* http/tests/security/cached-cross-origin-preloading-css-stylesheet.html:
* platform/mac-wk1/TestExpectations: Removed flaky expectation.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206769 268f45cc-cd09-0410-ab3c-d52691b4dbfc