[GTK] UIProcess crashes when using Japanese IM https://bugs.webkit.org/show_bug.cgi?id=163011
We have to reference the current GdkEventKey before we try process it
as later when the lambda body is reached the event could be already
freed.

Patch by Tomas Popela <tpopela@redhat.com> on 2016-10-10
Reviewed by Carlos Garcia Campos.

* UIProcess/API/gtk/WebKitWebViewBase.cpp:
(webkitWebViewBaseKeyPressEvent):
(webkitWebViewBaseKeyReleaseEvent):
* UIProcess/gtk/InputMethodFilter.h:
Use non-copyable Function so we can use WTFMove to pass the event to
lambda.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206985 268f45cc-cd09-0410-ab3c-d52691b4dbfc