REGRESSION(r201171): Web Inspector: Timeline Recording playhead should always start immediately, not wait until first event https://bugs.webkit.org/show_bug.cgi?id=163583 <rdar://problem/28815882>
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-10-18
Reviewed by Timothy Hatcher.

* UserInterface/Views/TimelineRecordingContentView.js:
(WebInspector.TimelineRecordingContentView.prototype._startUpdatingCurrentTime):
Revert logic change introduced by r201171 with no explanation.
Whenever we get a start time we should use it, regardless of
of what the current time is, precisely because the backend is
informing us of the start time to use.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@207473 268f45cc-cd09-0410-ab3c-d52691b4dbfc