Modern media controls don't update their rendering correctly https://bugs.webkit.org/show_bug.cgi?id=163603 <rdar://problem/28826022>
Patch by Antoine Quint <graouts@apple.com> on 2016-10-18
Reviewed by Dean Jackson.

There are rendering issues when we use an absolutely-positioned element
as the top-most element in a media element's shadow root (webkit.org/b/163592).
Since we only need for that element to be positioned, we can use "position: relative"
instead, which removes the rendering issues.

* Modules/modern-media-controls/controls/media-controls.css:
(.media-controls):
(.media-controls,): Deleted.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@207488 268f45cc-cd09-0410-ab3c-d52691b4dbfc