Backdrop filter doesn't show if removed then re-added https://bugs.webkit.org/show_bug.cgi?id=163860
Reviewed by Dan Bernstein.

Source/WebCore:

When a backdrop filter is removed the re-added, updateBackdropFiltersRect() never
runs the second time because m_backdropFiltersRect doesn't change. However, we need
to run that code to size and position the newly re-created backdrop layer, so run
it explicitly if we just created the backdrop layer. This is similar to how
updateContentsImage() calls updateContentsRects().

Test: css3/filters/backdrop/add-remove-add-backdrop-filter.html

* platform/graphics/ca/GraphicsLayerCA.cpp:
(WebCore::GraphicsLayerCA::setBackdropFilters):
(WebCore::GraphicsLayerCA::updateBackdropFilters):

LayoutTests:

* css3/filters/backdrop/add-remove-add-backdrop-filter-expected.html: Added.
* css3/filters/backdrop/add-remove-add-backdrop-filter.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@207722 268f45cc-cd09-0410-ab3c-d52691b4dbfc