Error description code should be able to handle Symbol values. https://bugs.webkit.org/show_bug.cgi?id=164436 <rdar://problem/29115583>
Reviewed by Filip Pizlo and Saam Barati.

JSTests:

* stress/error-description-on-symbols-should-not-crash.js: Added.

Source/JavaScriptCore:

Previously, we try to toString() the Symbol value, resulting in it throwing an
exception in errorDescriptionForValue() which breaks the invariant that
errorDescriptionForValue() should not throw.

We fixed this by making errorDescriptionForValue() aware of the Symbol type, and
not so a toString() on Symbol values.  Also fixed notAFunctionSourceAppender()
to build a nicer message for Symbol values.

* runtime/ExceptionHelpers.cpp:
(JSC::errorDescriptionForValue):
(JSC::notAFunctionSourceAppender):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208410 268f45cc-cd09-0410-ab3c-d52691b4dbfc