REGRESSION: Crashes in StringImpl destructor during GC when clearing the HasOwnPropertyCache https://bugs.webkit.org/show_bug.cgi?id=164433
Patch by Saam Barati <sbarati@apple.com> on 2016-11-08
Reviewed by Mark Lam.

Clearing the HasOwnPropertyCache will call deref() on the StringImpls
in the cache. We were doing this from the collector thread, which is
not allowed. It must be done from the mutator thread. We now clear the
cache in Heap::finalize() which happens before the mutator begins
executing JS after a collection happens.

* heap/Heap.cpp:
(JSC::Heap::collectInThread):
(JSC::Heap::finalize):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208426 268f45cc-cd09-0410-ab3c-d52691b4dbfc