We recursively grab a lock in the DFGBytecodeParser causing us to deadlock https://bugs.webkit.org/show_bug.cgi?id=164650
Reviewed by Geoffrey Garen.

JSTests:

* stress/dont-dead-lock-put-by-val-as-put-by-id.js: Added.
(ident):
(let.o.set foo):
(foo):

Source/JavaScriptCore:

Some code was incorrectly holding a lock when recursively calling
back into the bytecode parser's via inlining a put_by_val as a put_by_id.
This can cause a deadlock if the inlinee CodeBlock is something we're
already holding a lock for. I've changed the range of the lock holder
to be as narrow as possible.

* dfg/DFGByteCodeParser.cpp:
(JSC::DFG::ByteCodeParser::parseBlock):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208614 268f45cc-cd09-0410-ab3c-d52691b4dbfc