Web Inspector: inspector/worker/debugger-pause.html fails on WebKit1 https://bugs.webkit.org/show_bug.cgi?id=164787
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-11-15
Reviewed by Timothy Hatcher.

Source/JavaScriptCore:

* inspector/agents/InspectorDebuggerAgent.cpp:
(Inspector::InspectorDebuggerAgent::cancelPauseOnNextStatement):
Clear this DebuggerAgent state when we resume.

LayoutTests:

* inspector/worker/debugger-pause.html:
Make this test work for WebKit1 where the VM is shared between the
page and inspector page. We need to be able to stop the Inspector's
evaluation, so that we can evaluate and pause on the page, and then
come back to the inspector afterwards.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208755 268f45cc-cd09-0410-ab3c-d52691b4dbfc