[JSC] Build broken for 32-bit x86 after r208306 with GCC 4.9 https://bugs.webkit.org/show_bug.cgi?id=164588
Reviewed by Mark Lam.

Provide assembly for executing the cpuid instruction when compiling
in PIC mode with the GCC 4.9 EBX on 32-bit x86.

Note that the values returned by cpuid here are not used. The purpose
of calling this instruction is to force the CPU to complete and sync
any buffered modifications on registers, memory or flags before
fetching and executing the next instruction.

* wtf/Atomics.h:
(WTF::x86_cpuid):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208806 268f45cc-cd09-0410-ab3c-d52691b4dbfc