Unsupported emoji are invisible https://bugs.webkit.org/show_bug.cgi?id=164944 <rdar://problem/28591608>
Reviewed by Dean Jackson.

Source/WebCore:

In WidthIterator, we explicitly skip characters which the OS has no font
to render with. However, for emoji, we should draw something to show that
there is missing content. Because we have nothing to draw, we can draw
the .notdef glyph (empty box, or "tofu").

Test: fast/text/emoji-draws.html

* platform/graphics/WidthIterator.cpp:
(WebCore::characterMustDrawSomething):
(WebCore::WidthIterator::advanceInternal):

LayoutTests:

* fast/text/emoji-draws-expected-mismatch.html: Added.
* fast/text/emoji-draws.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208894 268f45cc-cd09-0410-ab3c-d52691b4dbfc