[Modern Media Controls] Add support for right-to-left layouts https://bugs.webkit.org/show_bug.cgi?id=165229
Patch by Antoine Quint <graouts@apple.com> on 2016-11-30
Reviewed by Dean Jackson.

Source/WebCore:

We implement the "usesLTRUserInterfaceLayoutDirection" property which is set by HTMLMediaElement
and set a CSS class when the layout is left-to-right, flipping the fullscreen volume slider in the
case that it would not be present (right-to-left).

Test: media/modern-media-controls/media-controller/media-controller-fullscreen-ltr.html

* Modules/modern-media-controls/controls/macos-fullscreen-media-controls.css:
(.media-controls.mac.fullscreen:not(.uses-ltr-user-interface-layout-direction) .volume.slider):
* Modules/modern-media-controls/controls/media-controls.js:
(MediaControls.prototype.get usesLTRUserInterfaceLayoutDirection):
(MediaControls.prototype.set usesLTRUserInterfaceLayoutDirection):
* Modules/modern-media-controls/media/media-controller.js:
(MediaController.prototype.set usesLTRUserInterfaceLayoutDirection):
(MediaController.prototype._updateControlsIfNeeded):

LayoutTests:

Add a test that toggles the layout direction and checks we correctly flip the volume slider.

* media/modern-media-controls/media-controller/media-controller-fullscreen-ltr-expected.txt: Added.
* media/modern-media-controls/media-controller/media-controller-fullscreen-ltr.html: Added.
* platform/ios-simulator/TestExpectations:
* platform/mac/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209170 268f45cc-cd09-0410-ab3c-d52691b4dbfc