[GTK] Force GDK backend to x11 when running with XvfbDriver https://bugs.webkit.org/show_bug.cgi?id=165612
Reviewed by Michael Catanzaro.

When the developer is running a Wayland session, GDK may get confused
and try to use its Wayland backend (which is higher priority) instead.
This ensures it will use the appropriate backend when running under
Xvfb.

* Scripts/webkitpy/port/gtk.py:
(GtkPort.setup_environ_for_server): when running with XvfbDriver,
force GDK_BACKEND to x11.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209606 268f45cc-cd09-0410-ab3c-d52691b4dbfc