Restore NSURLRequest's default time interval to match behavior before NSURLSession adoption https://bugs.webkit.org/show_bug.cgi?id=165821 <rdar://problem/28492939>
Reviewed by Brady Eidson.

Before adopting NSURLSession, iOS used CFURLConnection, not NSURLConnection.
iOS used to have a default timeout of INT_MAX and it now has a default timeout of 0, which means use the 
default NSURLRequest timeout, which is 60 seconds.  This is not enough for some slow mobile networks,
so we want to match behavior of our CFURLConnection code here.

* platform/network/ResourceRequestBase.cpp:
Use INT_MAX as the default timeout of requests on iOS.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209776 268f45cc-cd09-0410-ab3c-d52691b4dbfc