Only send Messages::WebPageProxy::DidSaveToPageCache once when entering page cache. <https://webkit.org/b/165887>
Reviewed by Brady Eidson.

We were sending one of these IPC messages per frame when a page enters the page cache,
and we really only need one, so only do it for the main frame.

(This message is used to adopt the page cache on the UI process side, in case another
web process currently owns the cache. That happens in WebProcessPool::processDidCachePage().)

* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::didSaveToPageCache):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209851 268f45cc-cd09-0410-ab3c-d52691b4dbfc