Cloned CachedResource should not have an empty response https://bugs.webkit.org/show_bug.cgi?id=165947
Patch by Youenn Fablet <youenn@apple.com> on 2016-12-17
Reviewed by Sam Weinig.

Source/WebCore:

Test: http/tests/loading/reusing-cached-stylesheet-from-different-domain.html

* loader/cache/CachedResource.cpp:
(WebCore::CachedResource::setBodyDataFrom): Copying the response to the cloned resource.
Even though the issue is appearing for stylesheets only, it is safer to do that for all cloned ressources.

LayoutTests:

This new test is working as follow:
- A first frame is loaded. As part of it, it will load an absolute URL CSS stylesheet that contains a relative URL to import a CSS stylesheet.
This second stylesheet will make the background as green.
- Once done, a second frame is loaded (same HTML but from a different domain) with the same absolute URL CSS stylesheet.
The relative URL stylesheet should use the absolute URL of the CSS stylesheet as base.
If it is using the frame as base, the loaded CSS will change background to red.

* http/tests/loading/resources/cached-stylesheet-from-different-domain-frame.css.php: Added.
* http/tests/loading/resources/cached-stylesheet-from-different-domain-frame.html: Added.
* http/tests/loading/resources/imported-stylesheet-varying-according-domain.css.php: Added.
* http/tests/loading/reusing-cached-stylesheet-from-different-domain-expected.html: Added.
* http/tests/loading/reusing-cached-stylesheet-from-different-domain.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209961 268f45cc-cd09-0410-ab3c-d52691b4dbfc